Storage: Addressing 64-bit windows warnings.
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@48851 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Storage/VMDK.cpp
@@ -1034,7 +1034,7 @@ out:
 static int vmdkReadGrainDirectory(PVMDKIMAGE pImage, PVMDKEXTENT pExtent)
 {
     int rc = VINF_SUCCESS;
-    unsigned i;
+    size_t i;
     uint32_t *pGDTmp, *pRGDTmp;
     size_t cbGD = pExtent->cGDEntries * sizeof(uint32_t);
 
@@ -2350,7 +2350,7 @@ static int vmdkDescriptorPrepare(PVMDKIMAGE pImage, uint64_t cbLimit,
      */
     size_t cbDescriptor = cbLimit ? cbLimit : 4 * _1K;
     char *pszDescriptor = (char *)RTMemAllocZ(cbDescriptor);
-    unsigned offDescriptor = 0;
+    size_t offDescriptor = 0;
 
     if (!pszDescriptor)
         return VERR_NO_MEMORY;
